+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 * Added fix for `thead` element without nested `tr` element.
 * Improved global inference in XmlProvider.
 * Write API for CsvProvider.
+* Remove Freebase provider.
 
 #### 2.1.1 - December 24 2014
 * Add SkipRows parameter to CsvProvider.